style(reporters): rename folder reporter/ -> reporters/, namespace the facade as Reporter.*
The folder holds many reporter implementations (tap/spec/dot/github/junit),
so the directory is plural; its index.ts is the singular general-purpose
Reporter facade. This fixes the filename<->identifier mismatch of a singular
`reporter/` folder yielding a plural `Reporters` namespace.

  import * as Reporter from ".../reporters/index.ts"
  createReporters -> Reporter.create
  reportRunStart  -> Reporter.runStart
  reportTestEnd   -> Reporter.testEnd
  reportRunEnd    -> Reporter.runEnd

No collision: the Reporter *interface* lives in reporters/types.ts and is
imported only by the reporter implementations, never by the facade
consumers. Dropped index.ts's unused type re-export (nothing imported the
types from the index; all use reporters/types.ts).
