+++ lib/reporters/dot.ts
